+++ Backend/Controller/product.js
@@ -9,6 +9,11 @@ import productModel from "../Model/productSchema.js";
  */
 export const postProduct = async (req, res) => {
     try {
+        const userId = req.user.id;
+        if(!userId) {
+            return res.status(404).json("There is no userId");
+        }
+
         const {title, description, category, image, location, type} = req.body;
         if(!title || !description || !location || !type) {
             return res.status(404).json("title, description, location and type are required");
@@ -21,7 +26,7 @@ export const postProduct = async (req, res) => {
 
         // save to database
         const savedItem = await productModel.create({
-            title, description, category, image, location, type
+            title, description, category, image, location, type, owner: userId
         })
         
         res.status(201).json({message: "Product Added Successfully", details: savedItem})
@@ -42,6 +47,9 @@ export const allPosts = async (req, res) => {
     try {
         const allPosts = await productModel.find();
 
+        if(allPosts.length < 1) {
+            return res.status(200).json([]);
+        }
         res.status(200).json(allPosts);
     } catch (error) {
         res.status(error.status || 500).json(error.message || "Something Went Wrong");
@@ -83,6 +91,18 @@ export const updateProduct = async (req, res) => {
     try {
         const {id} = req.params;
         const {status} = req.body;
+
+        // find the product
+        const product = await productModel.findById(id);
+
+        // check is access to do it for user
+        const userId = req.user.id;
+        if(!userId) {
+            return res.status(404).json("There is no userId");
+        }
+        if(product.owner.toString() !== userId) {
+            return res.status(403).json("You have no access to update this item")
+        }
         
         // check is status active or resolved
         if(status != 'active' && status != 'resolved') {
@@ -112,6 +132,17 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     try {
         const {id} = req.params;
+        const product = await productModel.findById(id);
+
+        // check access of user
+        const userId = req.user.id;
+        if(!userId) {
+            return res.status(404).json("There is no userId Provided");
+        }
+        if(product.owner.toString() !== userId) {
+            return res.status(404).json("You have no access to delete it")
+        }
+
         const deletedProduct = await productModel.findByIdAndDelete(id);
 
         if(!deletedProduct) {
